move tree builder to its own module
