+++ src/green/mod.rs
@@ -1,22 +1,24 @@
 //! The green tree is an immutable, persistent, atomically reference counted tree.
 
-mod token;
-mod node;
-mod element;
 mod builder;
 mod children;
+mod element;
+mod node;
+mod token;
+mod tree_builder;
 
 #[cfg(feature = "serde")]
 mod serde;
 
+pub(self) use self::element::{
+    pack_node_or_token, unpack_node_or_token, Element, FullAlignedElement, HalfAlignedElement,
+    PackedNodeOrToken,
+};
 #[doc(inline)]
 pub use self::{
-    builder::{Builder, Checkpoint, TreeBuilder},
+    builder::Builder,
     children::{Children, ChildrenWithOffsets},
     node::Node,
     token::Token,
-};
-pub(self) use element::{
-    pack_node_or_token, unpack_node_or_token, Element, FullAlignedElement, HalfAlignedElement,
-    PackedNodeOrToken,
+    tree_builder::{Checkpoint, TreeBuilder},
 };
